app/requirements.txt
attrs>=20.2.0
Click>=7.1.2
Flask>=1.1.2
Flask-Cors>=3.0.9
itsdangerous>=1.1.0
InowasFlopyAdapter>=1.5
Jinja2>=2.11.2
jsonschema>=3.2.0
MarkupSafe>=1.1.1
NumPy~=1.24.2
pandas>=1.5.3
pbr>=5.5.1
prometheus-flask-exporter>=0.18.1
pyrsistent>=0.17.3
pyproj==3.5.0
requests>=2.28.2
scikit-learn~=1.2.0
scipy>=1.5.3
six>=1.15.0
stevedore>=3.2.2
urllib3==1.26.15
virtualenv>=20.1.0
virtualenv-clone>=0.5.4
virtualenvwrapper>=4.8.4
Werkzeug>=1.0.1
